# [Python] Fix string replacement fields