Fixes #3649

This PR closely implements python's format string syntax specification.

see: https://docs.python.org/3.11/library/string.html#format-string-syntax

Only valid qualified identifiers are matched as replacement field names in order to reduce chance of false positives, even though python syntactically resolves all kinds of key names. _It appears to be an implementation detail instead of desired behaviour._

Replacement field termination is handled separately for each kind of quotation to ensure to correctly terminate replacement fields at the end of a string.

Format specs are handled more lazily following f-string implementation to better support time formats.

#### Benchmarks

Parsing performance is not effected noticeably.

Checked with 10k lines of differently complex templates:

```py
# 73ms -> 75ms
"From {} to {}"
```

```py
# 74ms -> 76ms
"From {0} to {0}"
```

```py
# 83ms -> 90ms
"From {0.prop[4]} to {0.prop[4]}"
```

```py
# 110ms -> 105ms
"From {0.prop[4]!r:{p}^9} to {0.prop[4]!r:{p}^9}"
```